use unsigned long long int instead of just long int to declare variable i in task 8

+++ 0x05-pointers_arrays_strings/8-print_array.c
@@ -12,11 +12,11 @@
  */
 void print_array(int *a, int n)
 {
-	long int i;
+	signed long long int i;
 
 	for (i = 0; i < n; i++)
 	{
-		printf("%d", a[i]);
+		printf("%ld", a[i]);
 		if (i != (n - 1))
 		{
 			_putchar(',');
